Add ignore_existing argument to org_create_assignment
Sometimes the Github API causes assignment creation to fail before all assignments have been created. I thought it would be nice to be able to re-run the same `org_create_assignment` command to pick up where assignment creation left off, so I added the option to run this function even if some of the requested repositories already exist. 

There may be reasons why this is a bad idea that I am just not aware of. Don't hesitate to reject this pull request if so.